pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tasman-etl"
version = "0.1.0"
description = "USAJOBS ETL: Chicago-focused data engineering roles → Postgres"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"                 # <-- string SPDX, avoids deprecation warning
authors = [{ name = "Ezra Chamba" }]

# Runtime deps
dependencies = [
  "requests>=2.32",                 # HTTP client
  "pydantic>=2.7",                  # parse/validate API payloads (v2 validators)
  "pydantic-settings>=2.0",         # parse/validate config (v2)
  "sqlalchemy>=2.0",                # DB engine / SQL
  "psycopg[binary,pool]>=3.2",      # Postgres driver (v3)
  "tenacity>=8.3",                  # retries w/ backoff+ jitter
  "structlog>=24.1",                # JSON logs to stdout
  "python-dotenv>=1.0",             # local .env dev
  "boto3>=1.34",                    # S3 bronze (optional)
  "great-expectations>=0.18.0",     # data quality gate
  "pandas>=1.5",
  "typing-extensions>=4.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2", "pytest-cov>=5.0",
  "testcontainers>=4.7.2",
  "mypy>=1.10", "ruff>=0.5.0",
  "pre-commit>=3.7",
  "types-requests",
  # "boto3-stubs[s3]>=1.34", 
  "types-urllib3",
]

# ----- tell setuptools where the package is -----
[tool.setuptools.package-dir]
"" = "src"                      # project root for packages is ./src

[tool.setuptools.packages.find]
where = ["src"]
include = ["tasman_etl*"]
exclude = ["tests*", "data*", "docker*", "infra*", "docs*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E","F","I","B","UP","PIE","C4","SIM"]
ignore = ["UP038"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"

[tool.pytest.ini_options]
filterwarnings = [
  # Suppress known Great Expectations warning about Validator-level `result_format` config not being persisted.
  # See: https://github.com/great-expectations/great_expectations/issues/7466
  # This is a temporary workaround until upstream resolves the issue.
  "ignore:`result_format` configured at the Validator-level will not be persisted.:UserWarning",
  "ignore:`Number` field should not be instantiated.*:marshmallow.warnings.ChangedInMarshmallow4Warning",
]
testpaths = ["tests"]
markers = [
  "integration: marks tests that perform database integration (deselect with '-m \"not integration\"')",
]
